+++ seeqret/migrations/initialize_database.py
@@ -9,6 +9,7 @@ from . import (
     db_v_003,
     db_v_004,
     db_v_005,
+    db_v_006,
 )
 
 
@@ -27,4 +28,6 @@ def init_db(vault_dir, user, email):
                 db_v_004.init_db_v_004(vault_dir)
             if cur < 5:
                 db_v_005.init_db_v_005(vault_dir)
+            if cur < 6:
+                db_v_006.init_db_v_006(vault_dir)
         cn.close()